# src/openapi.yaml
openapi: 3.0.3
info:
  title: Documenten API
  version: 1.2.0-rc4
  description: |
    Een API om een documentregistratiecomponent (DRC) te benaderen.

    In een documentregistratiecomponent worden INFORMATIEOBJECTen opgeslagen. Een
    INFORMATIEOBJECT is een digitaal document voorzien van meta-gegevens.
    INFORMATIEOBJECTen kunnen aan andere objecten zoals zaken en besluiten worden
    gerelateerd (maar dat hoeft niet) en kunnen gebruiksrechten hebben.

    GEBRUIKSRECHTEN leggen voorwaarden op aan het gebruik van het INFORMATIEOBJECT
    (buiten raadpleging). Deze GEBRUIKSRECHTEN worden niet door de API gevalideerd
    of gehandhaafd.

    De typering van INFORMATIEOBJECTen is in de Catalogi API (ZTC) ondergebracht in
    de vorm van INFORMATIEOBJECTTYPEn.

    **Uploaden van bestanden**

    Binnen deze API bestaan een aantal endpoints die binaire data ontvangen, al
    dan niet base64-encoded. Webservers moeten op deze endpoints een minimale
    request body size van 4.0 GiB ondersteunen. Dit omvat de JSON van de
    metadata EN de base64-encoded bestandsdata. Hou hierbij rekening met de
    overhead van base64, die ongeveer 33% bedraagt in worst-case scenario's. Dit
    betekent dat bij een limiet van 4GB het bestand maximaal ongeveer 3GB groot
    mag zijn.

    _Nieuw in 1.1.0_

    Bestanden kunnen groter zijn dan de minimale die door providers
    ondersteund moet worden. De consumer moet dan:

...
       meegestuurd wordt en de inhoud leeggelaten wordt.
...

...

...

...

...

...

...

...

...

...

...

...

...
paths:
...
  /api/v1/enkelvoudiginformatieobjecten/{uuid}/download:
    get:
      operationId: enkelvoudiginformatieobject_download
      description: Download de binaire data van het (ENKELVOUDIG) INFORMATIEOBJECT.
      summary: Download de binaire data van het (ENKELVOUDIG) INFORMATIEOBJECT.
      parameters:
      - in: query
        name: registratieOp
        schema:
          type: string
        description: Een datumtijd in ISO8601 formaat. De versie van het INFORMATIEOBJECT
          die qua `begin_registratie` het kortst hiervoor zit wordt opgehaald.
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
          description: Unieke resource identifier (UUID4)
          title: uuid
        required: true
      - in: query
        name: versie
        schema:
          type: integer
        description: Het (automatische) versienummer van het INFORMATIEOBJECT.
      tags:
      - enkelvoudiginformatieobjecten
      security:
      - JWT-Claims:
        - documenten.lezen
      responses:
        '200':
          headers:
            API-version:
              schema:
                type: string
              description: 'Geeft een specifieke API-versie aan in de context van
                een specifieke aanroep. Voorbeeld: 1.2.1.'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          description: OK
        '401':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Unauthorized
        '403':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Forbidden
        '404':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Not found
        '406':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Not acceptable
        '410':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Gone
        '415':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Unsupported media type
        '429':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Too many requests
        '500':
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/Fout'
          description: Internal server error
...
components:
...
  schemas:
...
    Fout:
      type: object
      description: Formaat van HTTP 4xx en 5xx fouten.
      properties:
        type:
          type: string
          description: URI referentie naar het type fout, bedoeld voor developers
          title: type
        code:
          type: string
          description: Systeemcode die het type fout aangeeft
          title: code
        title:
          type: string
          description: Generieke titel voor het type fout
          title: title
        status:
          type: integer
          description: De HTTP status code
          title: status
        detail:
          type: string
          description: Extra informatie bij de fout, indien beschikbaar
          title: detail
        instance:
          type: string
          description: URI met referentie naar dit specifiek voorkomen van de fout.
            Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.
          title: instance
      required:
      - code
      - detail
      - instance
      - status
      - title
...
  securitySchemes:
    JWT-Claims:
      type: http
      bearerFormat: JWT
      scheme: bearer
servers:
...